+++ test/chunking-form/samples/chunk-deshadowing-reassignment/_expected/amd/chunk2.js
@@ -1,10 +1,10 @@
 define(['exports', './chunk1.js', './chunk3.js'], function (exports, __chunk1_js, __chunk3_js) { 'use strict';
 
-	var x = x + 1;
+	var x = __chunk1_js.default + 1;
 
-	var y = x + 1;
+	var y = __chunk3_js.default + 1;
 
-	exports.x = x;
-	exports.y = y;
+	exports.default = x;
+	exports.default$1 = y;
 
 });