+++ test/chunking-form/samples/chunk-deshadowing-reassignment/_expected/cjs/chunk1.js
@@ -2,4 +2,4 @@
 
 var x = 42;
 
-exports.x = x;
+exports.default = x;