+++ test/chunking-form/samples/chunk-deshadowing-reassignment/_expected/es/chunk1.js
@@ -1,3 +1,3 @@
 var x = 42;
 
-export { x };
+export default x;