+++ test/chunking-form/samples/chunk-namespace-boundary/_expected/cjs/main2.js
@@ -1,0 +1,7 @@
+'use strict';
+
+var __chunk1_js = require('./chunk1.js');
+
+var main2 = __chunk1_js.default.map(d => d + 2);
+
+module.exports = main2;