+++ test/chunking-form/samples/chunk-namespace-boundary/main2.js
@@ -1,0 +1,2 @@
+import d from './shared';
+export default d.map(d => d + 2);