+++ test/chunking-form/samples/default-identifier-renaming/_expected/amd/main1.js
@@ -1,0 +1,7 @@
+define(['./chunk1.js'], function (__chunk1_js) { 'use strict';
+
+	var main1 = __chunk1_js.default.map(d => d + 1);
+
+	return main1;
+
+});