+++ test/chunking-form/samples/default-identifier-renaming/_expected/es/main1.js
@@ -1,0 +1,5 @@
+import d from './chunk1.js';
+
+var main1 = d.map(d => d + 1);
+
+export default main1;