+++ test/chunking-form/samples/default-identifier-renaming/shared.js
@@ -1,0 +1,2 @@
+const data = [1, 2, 3];
+export default data;