+++ test/chunking-form/samples/entrypoint-facade/_expected/cjs/main1.js
@@ -2,4 +2,4 @@
 
 var __chunk1_js = require('./chunk1.js');
 
-log(__chunk1_js.dep);
+__chunk1_js.default(__chunk1_js.default$1);