+++ test/form/samples/internal-conflict-resolution/_expected/iife.js
@@ -7,10 +7,10 @@
 		return bar;
 	}
 
-	function bar$2 () {
+	function bar$1 () {
 		alert( foo() );
 	}
 
-	bar$2();
+	bar$1();
 
 }());